# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Plataforma Web MOOC Académica
  description: Backend monolito modular en Go para la plataforma de cursos masivos abiertos en línea.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Entorno Docker Local
paths:
...
  /api/v1/auth/register:
    post:
      summary: Registrar nuevo estudiante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                full_name:
                  type: string
      responses:
        '200':
          description: Estudiante registrado exitosamente
  /api/v1/auth/verify-email:
    post:
      summary: Verificar correo electrónico con token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Correo verificado exitosamente
  /api/v1/auth/login:
    post:
      summary: Inicio de sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Autenticación exitosa con token Bearer retornado
  /api/v1/auth/logout:
    post:
      summary: Cierre de sesión y revocación de token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sesión cerrada exitosamente
  /api/v1/auth/forgot-password:
    post:
      summary: Solicitar restablecimiento de contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Instrucciones enviadas si el usuario existe
  /api/v1/auth/reset-password:
    post:
      summary: Restablecer contraseña con token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Contraseña actualizada exitosamente
  /api/v1/admin/users:
    get:
      summary: Listar usuarios (Solo Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista paginada de usuarios
  /api/v1/admin/teachers:
    post:
      summary: Crear nuevo profesor (Solo Admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
      responses:
        '200':
          description: Profesor creado exitosamente
  /api/v1/admin/users/{id}/status:
    patch:
      summary: Cambiar estado de usuario (Solo Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, suspended, unverified]
      responses:
        '200':
          description: Estado de usuario actualizado
  /api/v1/admin/audit-logs:
    get:
      summary: Consultar registros de auditoría (Solo Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Registros de auditoría paginados
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer